- fixed compiler warnings; use macros from <inttypes.h> when printing    int16_t and cast sizeof() operator to int16_t.

+++ padProtoHost.c
@@ -12,6 +12,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <inttypes.h>
 
 static int verbose    = 0;
 
@@ -41,7 +42,7 @@ int16_t tmp;
 	printf("     Type: 0x%02x\n",           (uint8_t)p->type);
 	printf("  Channel: %i\n",               p->chnl);
 	tmp = ntohs(p->nBytes);
-	printf("   nBytes: %i [payload %i]\n",  tmp, tmp-sizeof(*p));
+	printf("   nBytes: %"PRIi16" [payload %"PRIi16"]\n",  tmp, tmp-(int16_t)sizeof(*p));
 	printf("      XID: 0x%08x\n",           ntohl(p->xid)); 
 	tmp = ntohs(p->status);
 	printf("   status: %i (%s)\n",          tmp, strerror(-tmp));
